move process-event under src/worker
This one is sort of weird since it gets called in server.ts
