+++ src/types.ts
@@ -2,7 +2,6 @@ import ClickHouse from '@posthog/clickhouse'
 import { PluginAttachment, PluginConfigSchema, PluginEvent, Properties } from '@posthog/plugin-scaffold'
 import { Pool as GenericPool } from 'generic-pool'
 import { StatsD } from 'hot-shots'
-import { EventsProcessor } from 'ingestion/process-event'
 import { Redis } from 'ioredis'
 import { Kafka, Producer } from 'kafkajs'
 import { DateTime } from 'luxon'
@@ -10,6 +9,7 @@ import { Pool } from 'pg'
 import { VM } from 'vm2'
 
 import { DB } from './db'
+import { EventsProcessor } from './worker/ingestion/process-event'
 import { LazyPluginVM } from './worker/vm/lazy'
 
 export enum LogLevel {
